+++ script/tool/test/java_test_command_test.dart
@@ -43,8 +43,8 @@ void main() {
       final Directory plugin = createFakePlugin(
         'plugin1',
         packagesDir,
-        platformSupport: <String, PlatformSupport>{
-          kPlatformAndroid: PlatformSupport.inline
+        platformSupport: <String, PlatformDetails>{
+          kPlatformAndroid: const PlatformDetails(PlatformSupport.inline),
         },
         extraFiles: <String>[
           'example/android/gradlew',
@@ -73,8 +73,8 @@ void main() {
       final Directory plugin = createFakePlugin(
         'plugin1',
         packagesDir,
-        platformSupport: <String, PlatformSupport>{
-          kPlatformAndroid: PlatformSupport.inline
+        platformSupport: <String, PlatformDetails>{
+          kPlatformAndroid: const PlatformDetails(PlatformSupport.inline),
         },
         extraFiles: <String>[
           'example/android/gradlew',
@@ -103,8 +103,8 @@ void main() {
       createFakePlugin(
         'plugin1',
         packagesDir,
-        platformSupport: <String, PlatformSupport>{
-          kPlatformAndroid: PlatformSupport.inline
+        platformSupport: <String, PlatformDetails>{
+          kPlatformAndroid: const PlatformDetails(PlatformSupport.inline)
         },
         extraFiles: <String>[
           'example/android/app/src/test/example_test.java',
@@ -133,8 +133,8 @@ void main() {
       final Directory pluginDir = createFakePlugin(
         'plugin1',
         packagesDir,
-        platformSupport: <String, PlatformSupport>{
-          kPlatformAndroid: PlatformSupport.inline
+        platformSupport: <String, PlatformDetails>{
+          kPlatformAndroid: const PlatformDetails(PlatformSupport.inline)
         },
         extraFiles: <String>[
           'example/android/gradlew',